requirements.txt
pickleshare==0.7.5
numpy==1.18.1
Pillow==7.0.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
torch==1.4.0
torchvision==0.5.0